+++ __main__.py
@@ -25,9 +25,7 @@ if __name__ == '__main__':
 
     try:
         context = allInput.get('context')
-        if (functionName == 'loadRemoteEvent'):
-            result = transformer.loadRemoteEvent(event)
-        elif (functionName == 'loadAndUpdateRemoteEvent'):
+        if (functionName == 'loadAndUpdateRemoteEvent'):
             result = transformer.loadAndUpdateRemoteEvent(event, context)
         elif (functionName == 'loadNestedEvent'):
             result = transformer.loadNestedEvent(event, context)